Fixed remapping quirks due to RemappingClassAdapter calling map functions with invalid access parameter.
